Don't fail tests after after 5.9 seconds.
5 seconds after the first packet, our IPv6 router ND cache
entries go into PROBE, and because the retran timer is mistakenly
set to 300ms instead of 300s, after 5.9s they go into FAILED,
making it impossible to send any further packets.

This is likely the cause of the UID routing failures observed on
the continuous build servers, so don't skip them any more.

Change-Id: I4cfa81c901bf6f63861dd1ace7bc54c3067e2f37