Test for trace events used by last-mile debugging
Add a test that verifies that all of the events
needed by last-mile debugging a) exist, and
b) only use whitelisted pointer formats.

Along the way: add some patches to get the
marlin kernel to compile for UML, with a) tracing
support, and b) cfg80211 support.

Bug: 30142966
Test: rm -f $NET_TEST/net_test.rootfs* && \
      for kernel in \
          $ANGLER_KERNEL $BULLHEAD_KERNEL $MARLIN_KERNEL $SHAMU_KERNEL; do \
 	     echo "*** START CLEAN: $kernel" && \
	     cd $kernel && \
	     make mrproper && \
	     git reset --hard && \
	     git clean -d --force --quiet && \
	     $NET_TEST/run_net_test.sh --builder trace_event_test.py && \
	     echo "*** START DIRTY: $kernel" && \
	     $NET_TEST/run_net_test.sh --builder trace_event_test.py; \
	     if (( $? )); then \
	         echo "*** FAIL: $kernel"; \
		 break; \
	     fi \
      done >& $LOG_FILE

Test note: The output of the test commands above must be
directed to a file. If we run the above test commands with
output directed to a terminal, then the second kernel build
fails due to write() on the PTY returning EAGAIN. (I assume
this happens because UML has somehow left the PTY in a bad
state.)

Change-Id: Ic6c6ecb7b42a9d82667d6b61619d96e583aeb7cf