net_test: Add test for RFC7559 router solicitation backoff
resilient_rs_test tries to verify that RS backoff follows an exponential
schedule. This test should fail with high probability if RFC7559 is not
implemented and enabled in the kernel. Note that due to a minimum delay
of 1 second, this test may require up to 10 seconds to gather sufficient
evidence of exponential backoff.

Test: net_test passes
Bug: 33620395
Change-Id: I1ba6ab384d4c5654e3268842a2933375ad0e69f9