Fix a race in ping6_test.py.
ping6_test starts one PingReplyThread per test network, but
then immediately starts the tests without waiting for the
reply threads to be started. If the tests run before the
reply threads are started, they will fail with EAGAIN.

This CL makes the test wait a configurable time (currently 2
seconds) for all the threads to start.

Bug: 31119353
Change-Id: I488c9fb99e874df1ff3c929dc7719f362208fdf7