Add 10% of slack when expecting NUD_PROBE notifications.
This makes NeighbourTest.testNotifications pass on net-next. I
don't know what changed; the test mostly passed in July and no
longer passes now, but a few reverts to net/core/neighbour.c did
not reveal a smoking gun. It could be something unrelated such
as a UML timer change, or perhaps additional jitter added to
the notifications.

In any case, failing after waiting exactly 1000ms for a 1000ms
timer seems brittle - in fact, it's surprising that the test
was as reliable as it was. It looks like we already have a 20%
fudge factor when expecting NUD_STALE notifications.

Change-Id: I97da1a51df80e2c51897bb8366bc35a0a8ffc88b