Update patch set 2
Patch Set 2:

(10 comments)

A good start! Comments inline.

This code will be most useful if we can hook it up into the actual packet send and receive code and verify that BPF code that we program actually does what we want. If there a way we can do this? For example, can we set a socket input filter, then pretend we received a packet, and check that our input filter does something useful with it?

Patch-set: 2
Label: Verified=0