Add test for bpf helper function
Add two more test for bpf helper function getSocketCookie and
getSocketUid, the test only verifies if the helper function is
implemented or not. It cannot check if the result of the helper function
is correct since the socket filed in sk_buff is missing when socket
filter is applied. Further test need to be implemented when the xt_ebpf
module is avalaible.

Test: all_tests.sh passes on android-4.4 or above

Change-Id: Iea8f51720f44068393545787cc256c43576d1309