Update patch set 2
Patch Set 2:

(1 comment)

I'm not a fan of adding kernel patches to net-test, because it has to run on such a wide variety of kernels. Currently we have net-test running and passing on kernels all the way from 3.10 all to 4.10-ish (net-next), on upstream, generic android trees, and devices of many different vintages.

The approach that we've taken for all current device kernels (and the one that is documented at https://source.android.com/devices/tech/config/kernel_network_tests.html ) is that if a device-specific kernel change breaks compilation for ARCH=um, then it's a bug in that device kernel and needs to be fixed in that kernel.

For the wifi_power_mgmt patch, I think it's probably best to get it into the marlin kernel only (and the kernels of any other devices that share the same vintage; we have lots of continuous tests running, and if the tests fail then we'll cause build break emails).

The arch_timer patch is sort of borderline, because it exists upstream, albeit only for arm and arm64. It might be worth trying to get that into the upstream UM kernel. Not sure what the maintainers would say though.

How extensive is the breakage if arch_counter_get_cntpct is not defined? Is it just that the marlin kernel breaks in a couple of places, or is there breakage on lots of devices in lots of places?

Patch-set: 2
Label: Verified=0