Update patch set 2
Patch Set 2:

(1 comment)

Fair point about the patching -- I had some reservations myself. And thanks for the pointer to the kernel test docs.

Not having arch_counter_get_cntpct() only causes a problem in one place. (To be clear: there is more than one call-site for arch_counter_get_cntpct(). But only one of the call-sites is relevant with our choice of config options.)

I'll upload some marlin-kernel patches so you can chime in on whether that approach is sensible. I'll circle back on this CL when those are ready.

Patch-set: 2