Update patch set 1
Patch Set 1:

Unless this test can check for correct behaviour, it seems like it would be best not to merge it for now, and instead use it to do test-driven development of the test patches.

If https://android-review.googlesource.com/#/c/317697/ were in the kernel, can this test be made more detailed? If so, I would suggest making this CL the unit test for that CL, and merge them around the same time.

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0