Update patch set 1
Patch Set 1:

(1 comment)

The name "qstruct" SGTM, but how common is the cstruct module that we are conflicting here? I took a quick look and it doesn't look like it's packaged for ubuntu at least. So if the risk of conflict is not high, I'd prefer to keep the name as is in order not to muddy the blame/history of these files.

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>