Update patch set 2
Patch Set 2: Code-Review-1

I don't think we should do this. I understand that the naming is somewhat unfortunate, but this CL is changing a fair bit of code and affecting blame/history. If we're trying to make it easier for new developers to work in this code, then perhaps a judicious set of comments in the code, or text added to the README, might help?

Patch-set: 2
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1