Test all* algorithm combinations.
This tests all supported combinations of encryption and authentication
algorithms, as defined in IpSecAlgorithm.java. In particular, this test
does not cover any AEAD algorithms. (Those are not supported yet, and
the netlink attr type is different.)

For each cipher + hmac, we also test all valid key lengths and min/max
truncation lengths.

Bug: 62071755

Test: all_tests.sh passing on device kernels

Change-Id: I5aa5418a97eab3961ea495785671e65e0c0d7df7