Update patch set 3
Patch Set 3: Code-Review-1

This rule is representative of what happens on actual devices, so in order to change it here we should change it on devices as well, otherwise our tests won't match reality.

But I don't think this rule is unneeded. Off the top of my head, one thing that it's needed for is to ensure that packets sent by a socket that was connected on a now-disconnected network correctly get ENETUNREACH instead of being sent on our current default network with a source IP address that's not correct for network we're currently on.

Patch-set: 3
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1