Enforce 'yapf --style=chromium' in kernel tests.
This enforces chromium style on modified Python files of subsequent CLs.
The script is branched from tools/test/connectivity because submitting
patches to that project requires passing a number of unrelated preupload
hooks.

Chromium style is similar to Google style, but uses 2-space indents.

Note that this is really only concerned with whitespace, and lint errors
are a separate matter.

Test: the hook enforced style on yapf_checker.py

Change-Id: I97407fa6887ecd27d2edadfce9b450b723ab6f3f