Test verify qtaguid not drop packet without socket.
Add a test to verify that the xt_qtaguid module will not drop a packet
if the cooresponding socket is closed. Since with the help of sk->uid
added inside sk_buff, qtaguid no longer depend on the socket file to
determine the uid of a packet. So the uid owner match should be
functional even if the socket is closed.

Bug: 37524657
Test: qtaguid_test.py should pass on all kernels
Change-Id: I3c935a0def393b29e43d80a2949bb1d4ead66155