Update patch set 1
Patch Set 1:

Does this test pass on any current kernel?

I think the goal is that it fails everywhere. What we're trying to do is find something that fails 100% of the time on current kernels, but will be fixed when we move xt_qtaguid to using sk_uid.

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>