Update patch set 1
Patch Set 1:

(1 comment)

I'm still not in favor of this. I'd rather have the parameters passed to the function. If we're going to make a simplifying equivalence about the addresses in the selector, I'd rather just take None and handle that inside the tunnel mode creation. it would negate half of the parameter "savings" without introducing yet another new paradigm into our function signatures.

Alternatively, maybe we should start creating "helper" classes for things like the algorithms and passing those in to AddMinimalSaInfo and ditch the passing of packed structs (which may or may not be pre-prended with NLATTR headers)?

Patch-set: 1
Reviewer: Gerrit User 1060151 <1060151@85c56323-6fa9-3386-8a01-6480fb634889>