Update patch set 9
Patch Set 9: Code-Review+1

(13 comments)

Mostly LGTM with a few style comments.
The one substantive comment is: what's causing integrity_failed to go up? Is it that the SPI is incorrect? (If so, should that even happen? Seems like it should not find the SA at all instead.)  Should we put back the code that tampers with the packet?

Patch-set: 9
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1