[DO NOT SUBMIT] Test multiple VTI tunnels at the same time.
XfrmVtiTest now creates one IPv4 and one IPv6 tunnel per netid.
Similar to MultinetworkBaseTest, the tunnels are created in
setUpClass and torn down in tearDownClass, so all the tests are
always run with all the VTIs present.

This provides the basic infrastructure to perform more in-depth
testing of the VTI code in a realistic environment.

Bug: 70371070
Test: xfrm tests pass on android-4.9
Change-Id: I67b3f28f3357088ad1c51ddc5f7b80c131ed651c