Make incoming mark tests pass on device.
The tests set up routing in an overly simplistic way, with all
mark-based rules all having a mask of 0xffffffff. This does not
work on device, where inbound marks all contain permission bits
and the accounting bit.

Fix this by making all mark rules have a mask of 0xffff. While
not fully representative of what netd does, it's enough to get
the test to pass on device.

Bug: 36510518
Test: MarkTest and TCPAcceptTest pass on device
Change-Id: Ib6435a85b20e26c700d390455ec32c31b67f56de