Don't require that we be the only sockets in the system.
Currently, CheckSockStatFile compares the expected values against
the last line in /proc/net/{icmp,tcp,udp}{,6} and fails if they
don't match. This means that the test only passes if the socket
being looked for is the last in the file.

Instead, check the whole file and make the test pass if the
socket being looked for appears anywhere in the output.

Bug: 36510518
Test: ping6_test.py passes on device
Test: all_tests.sh passes on android-4.9
Change-Id: I89b12f73c40961b0ef66c4716b0dc7187f42d453