Add more context to decoding netlink attributes.
Netlink attributes are in general hard to interpret because the
NLA types are always enums that start with 1, and therefore their
meaning always depends on context.

Currently the only context we pass into the attribute decoding
method is the command name or attribute name. This is not
sufficient in some cases. For example, interpreting attributes
nested inside in IFLA_INFO_DATA attribute requires knowing at
least the IFLA_INFO_KIND attribute of the link, which is
nested at the same level as IFLA_INFO_DATA.

This change makes it so that when decoding nested attributes we
pass in a dictionary containing the attributes of the parent
attribute that we have decoded so far.

Bug: 70371070
Test: all tests except pf_key_test (expected) pass on ipsec-next-xfrmi2
Change-Id: I10f58d05ae553a2f85d1137641b098e36db1a147