Test To Show That SAs are Ambiguous With Marks
Because of the way that marks are implemented in
XFRM, two SAs that differ by only the mark cannot
be reliably accessed. This test demonstrates the
fundamental problem, which we cannot work around
and also codifies for the future that an SA scheme
that relies on operational SAs with two different
levels of mark matching is fundamentally broken.

Bug: 71645364
Test: run_net_test.sh xfrm_test.py
Change-Id: Id130bd8ff72808d203d44b523093b013b674fea1