anycast_test.py: make sure CloseFileDescriptorThread finished
The original implementation close fd in a separate thread,
which might cause the check in this thread run first before
the fd close in CloseFileDescriptorThread.

This change uses thread.join to make sure fd close finished
before the later check if there is no hang problem.
And use thread.isActive to check if there is the hang problem.

Test: run vts-kernel -m VtsKernelNetTest with 4.9 and 4.14 kernel

There will be 1 or more failed in 5 iterations,
after this change all will passed in 10 iterations.

Change-Id: I24341667c9a889d79d38a9eb1d88d05325d7ffff
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>