Update patch set 1
Patch Set 1:

YongQin, you're saying that this test fails about 1 out of 5 times without this patch. I can't reproduce that - it passed 100 times in a row on taimen. Can you explain how it fails?

On the change itself - it seems to me that the change is roughly equivalent to changing the time.sleep(0.1) to time.sleep(0.2). Did you consider doing that instead?

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>