Update patch set 22
Patch Set 22:

(17 comments)

Mostly LGTM but I tihnk we shouldn't tear down state in the middle of tests like this, since if we do that we can't be confident that input, output, ICMP etc. work on the same tunnel objects - which is ultimately what we're trying to test.

Patch-set: 22
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>