Update patch set 1
Patch Set 1: Code-Review-1

I don't think we should just remove this code. The intent of the test is to ensure that if we set a socket policy that requires an SA, and there is no SA, then nothing is sent and an error is returned.

What is the behaviour when a userspace key manager exists? That the packet is dropped, an XFRM_MSG_ACQUIRE message is sent, and an XFRM state in state XFRM_STATE_ACQ is created? If so, then instead of deleting the code that checks for EAGAIN, we should change it to check for either of two outcomes:

Outcome #1: sendto returns EAGAIN and no packet is sent
Outcome #2: sendto returns 0, and an XFRM state in state XFRM_STATE_ACQ is created.

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1