Update patch set 1
Patch Set 1:

> about #Comment1(Line:136), at the begining of the test case, FLUSHSA(I have received such msg--flush state proto ipsec-any) &FLUSHPOLICY are called, so there is no SAinfo & Policyinfo. if wifi calling is on, epdg tunnel is still cleared because FLUSHSA&FLUSHPOLICY is executed

It's true that many of these tests flush all state, but that's not a good state of affairs. Actually, it's bug that we would like to fix. If possible, we should not add new code that depends on this property.

> about #comment2(Line:149), this code is necessary. If there is a user space key manager existing, there will be a XFRM_MSG_ACQUIRED netlink msg sending,so dumpSAinfo will have a acquired state info, len(dump)==1.sendto also poped up EAGAIN. That is what MTK platform real happened on a real device, I have tested it. XFRM_MSG_UPDSA should be instead of XFRM_MSG_ADDSA when there is a key manager existing. XFRM_MSG_UPDSA will update state info to match the socket policy. I think that can also test what the test case wanted

Ok, so what about the following?

1. Apply the policy.
2. Call sendto() and check that it returns EAGAIN.
3. Delete the embryonic SA that we just created.
4. Continue with the existing code (line 137-).

If #3 is more than a 5-10 lines of code it might need to be in a separate method.

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>