Update patch set 21
Patch Set 21:

(5 comments)

Code generally LGTM (minus the code duplication), but it's not immediately clear why we need this test. Are we just checking that when an XFRMI or VTI tunnels are configured on the host, that native packets continue to be sent and received? If so, we should say that in the commit message and the comments.

Patch-set: 21
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>