Update patch set 3
Patch Set 3: Code-Review+1

(1 comment)

Code LGTM, thanks!

Hung-Ying, is it appropriate to mention a particular SOC vendor (MediaTek) in the commit message? This problem is not specific to MediaTek but can occur on any platform where an IPsec key manager application is running.

If it's not appropriate to mention a particular SOC vendor in the commit message, then the message can simply be reworded to "On a platform where an IPsec key manager is running...".

Patch-set: 3
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1