Add tests for IPsec tunnel mode over loopback
This commit adds a test to ensure that IPsec tunnel mode over loopback
works. In old kernels, the IPsec secpath is not cleared before sending
the packet back into the receive path via loopback. As such, the
outbound states are still in the secpath, causing inbound policy match
to fail with XFRMINNOPOLS

This patch tests to ensure that the secpath_reset(skb) is called in
loopback_xmit.

Test: New loopback test added into xfrm_tunnel_test
Change-Id: I8057e3c0934ae2ef281efd99620c4f1f2b49e457