DO NOT MERGE: improve xfrm net test
in kernel_net_test/xfrm_tunnel_test.XfrmVtiTest

the testcases write packets to the tunnel, then check the count of rx/tx
packets number on the interface.

however, the packets are processed *asynchronously* in kernel.

the check of rx/tx packets maybe failed.

a delay is added here, then the testcase will pass in most cases, just a
workaround.

Test: run vts/kernel_net_tests
Bug: 117532237
Signed-off-by: Xu YiPing <xuyiping@hisilicon.com>