Update patch set 2
Patch Set 2:

(7 comments)

Generally speaking I'm not a fan of inline python like this since it's not easy to syntax check and maintain, but if we can make this code a bit simpler I can live with it.

I also can't get away from thinking that we're hacking around a problem here. If the only thing that breaks when we don't have enough randomness is /proc/net/icmp6, then shouldn't we fix that instead?

Patch-set: 2
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>