net_test: workaround for 3.18 and 4.4 late urandom init
By default it initializes 60 seconds after boot (if there's enough
entropy at that point in time), which is too late (we don't want
to simply wait 60 seconds, but we do want it initialized prior
to running any tests).

We also switch from hex (base 16) to base64 encoding of the entropy
string just to make the kernel command line shorter.

Change-Id: Ia4924576c550b2858a938873baef18893550fa04