Update patch set 1
Patch Set 1:

> > > We already fixed this in https://android-review.googlesource.com/c/kernel/tests/+/813793/ . We should ensure that the test is run by that change instead.
> > 
> > The fix you mentioned is already distributed in VTS9.0 R3, but the kernel with MPTCP fails with this VTS version.
> > 
> > I confirmed that if we exclude MPTCP from our build, VtsKernelNetTest is passed.
> 
> I asked multipath-tcp.org about this delay, and there is already a discussion about this.
> According to the discussion, it is due to DAD procedure within MPTCP fullmesh module. It seems to be the duplicate address-timer takes a refcount on the IPv6-address, preventing it from getting closed.

Thanks for the follow-up.

Out of curiosity, on which interface is DAD being performed? Can we fix this by disabling DAD on the interface?

Failing that: can we base the fix for this issue on https://android-review.googlesource.com/c/kernel/tests/+/813793 ? Increase the 0.5 second timer to something higher (2 seconds?) and add a comment explaining that this is due to MPTCP (and linking to the message on the mptcp mailing list where this is explained, if you have such a message)

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>