Update patch set 1
Patch Set 1:

As discussed in https://android-review.googlesource.com/c/kernel/tests/+/839290, I think the waiting timer should be increased to 2 seconds or more.

For kernels that MPTCP is ported, it take more than 0.5 sec to close tun interface, because of DAD procedure within MPTCP fullmesh module. It seems to be the duplicate address-timer takes a refcount on the IPv6-address, preventing it from getting closed.

Patch-set: 1
CC: Gerrit User 1272272 <1272272@85c56323-6fa9-3386-8a01-6480fb634889>