Update patch set 3
Patch Set 3: Presubmit-Ready+1

Once the manifest changes go in to version the kernel tests, this should be good to go.

I'd prefer not to break tests if possible, and if it's only a day or two, I think we should wait.

Patch-set: 3
Reviewer: Gerrit User 1221725 <1221725@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1