Update patch set 2
Patch Set 2:

> Braindumping a few thoughts:
> 
> Perhaps for the affected tests, we can drop neighbor solicitations on all the relevant interfaces (either via iptables, or disabling autoconf)

It does seem fine to lengthen the NUD timers for the XFRM tests only (that's cleaner than using iptables). But we shouldn't lengthen them for other tests.

Iris, can you modify the CL to change the NUD timers only in the XFRM tests that are failing? It should be sufficient to call SetSysctl in the setUpClass method of xfrm_tunnel_test.

Patch-set: 2
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>