Update patch set 4
Patch Set 4: Code-Review+1

> Patch Set 4:
> 
> I'm not sure how to permit the automerger...

Sorry, I really meant TreeHugger here.  It was permitted at All-Projects level, if the goal is to prevent everyone else other than the permitted bots and android-core-networking, the change looks fine to me.

Note that if your intention is to also restrict "Autosubmit" to the new group only (in other words, someone from the group has to approve submission to the two branches), you should add exclusiveGroupPermissions = label-Autosubmit and label-Autosubmit = 0..+1 group mdb/android-core-networking to the two branches.

Also, could you please amend the commit message so that it says something like "Restrict submit access to android-core-networking" instead of the generic "Review access change" so that we know the removal was intentional when someone else reads this in the future?

Patch-set: 4
Reviewer: Gerrit User 1086523 <1086523@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1