Update patch set 1
Patch Set 1:

afaict android-common-4.4 on the master branch is not cuttlefish compatible.

This is because cuttlefish is a shipping api level Q device with a injected 4.4 kernel which is not eBpf capable.

This results in netd enabling ebpf support and then crashing the framework. 

04-16 23:16:23.038  1437  1437 I netd    : trafficSwapActiveStatsMap() -> ServiceSpecificException(95, "[Operation not supported on transport endpoint] : This device doesn't have eBPF support") <0.01ms>
--------- beginning of crash
04-16 23:16:23.039  1666  1773 E AndroidRuntime: *** FATAL EXCEPTION IN SYSTEM PROCESS: NetworkStats
04-16 23:16:23.039  1666  1773 E AndroidRuntime: android.os.ServiceSpecificException: [Operation not supported on transport endpoint] : This device doesn't have eBPF support (code 95)

Patch-set: 1
Reviewer: Gerrit User 1003247 <1003247@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0