Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> What does it even mean that the tests are being migrated to Java?
> 
> Don't these tests require root and a bunch of capabilities?

The goal is to not use the vts python runner any more. KernelNetTest.java is a temp solution that uses a java test to run a binary on the device side. We will have a more generic runner to do that.
The root requirement will still be applied through RootTargetPreparer:
https://cs.corp.google.com/android/test/vts/testcases/host/kernel_net_test/AndroidTest.xml?l=21

Patch-set: 6
Reviewer: Gerrit User 1088076 <1088076@85c56323-6fa9-3386-8a01-6480fb634889>