Update patch set 4
Patch Set 4:

(1 comment)

> Benedict, do you remember what the actual requirements were on this test?
As I remember, the original requirements here was P+ for everything minus XFRM-I tests, and Q+ for XFRM-I tests.

> I think we're still running the Pixel 2 (4.4 kernel, launched on O) - which didn't pick up a Q era kernel - against the old/branched net tests: ie. 
IIRC, we did backport the changes necessary to pass this on 3.18, 4.4, 4.9 and 4.14 to the common kernels (non-release branch), and pixel picked this up all the way back to pixel 1.

> Basically I think this is trying to solve bug 134443647 'Add --api_level Q or something to kernel net tests' so that we could add a @SkipUnlessKernel(Q+) clause.
Realistically, this would be the ideal case; I think we didn't have the ability to do this at the time the tests were written, but if we can enforce based on launch version, that would be amazing.

Patch-set: 4
Reviewer: Gerrit User 1221725 <1221725@85c56323-6fa9-3386-8a01-6480fb634889>