Update patch set 1
Patch Set 1: Code-Review-1

I think it would be better to run in the root namespace if possible, because the tests will be more realistic. I only recently found a flake in our tests (which pointed to a bug in our stack) due to a test being flaky because our devices usually have forwarding turned on.

Maybe we can run in a separate namespace iff there is an active adb-over-tcp connection? We could assume that the port is the default (5555), use a SockDiag object to determine all the connections with sport 5555, and check if the owning pid is adbd.

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1