net-test: Check that xt_qtaguid is truly gone.
VTS has required NETFILTER_XT_MATCH_QTAGUID=n for a while, but
it doesn't actually check that the code is not in the kernel.

To be precise:
- we switched to bpf in P
- xt_qtaguid was removed in Q

P is 4.4, 4.9, 4.14
Q is 4.9, 4.14, 4.19
R is 4.14, 4.19, 5.4

If we look at Android Common Kernel 4.14-P it does have xt_qtaguid.

Similarly VINTF for P-4.14 also permits QTAGUID.

As such minimum requirement is 4.19+ due to 4.14-P

Bug: 138428914
Test: all_tests.sh passes locally
Change-Id: Iabc4af24092d6d92c81b16dd305600a6785bf700